refactor: lib/network -> 각자의길
